+++ src/mapfile_parser/__init__.py
@@ -13,7 +13,7 @@ from . import utils as utils
 
 from .mapfile import MapFile as MapFile
 from .mapfile import Symbol as Symbol
-from .mapfile import File as File
+from .mapfile import Section as Section
 from .mapfile import Segment as Segment
 from .mapfile import FoundSymbolInfo as FoundSymbolInfo
 from .mapfile import SymbolComparisonInfo as SymbolComparisonInfo
@@ -22,3 +22,6 @@ from .mapfile import MapsComparisonInfo as MapsComparisonInfo
 from .progress_stats import ProgressStats as ProgressStats
 
 from . import frontends as frontends
+
+# Renamed types
+from .mapfile import Section as File