+++ src/rs/lib.rs
@@ -1,26 +1,30 @@
 /* SPDX-FileCopyrightText: © 2023-2025 Decompollaborate */
 /* SPDX-License-Identifier: MIT */
 
-mod file;
 mod found_symbol_info;
 mod mapfile;
 mod maps_comparison_info;
 mod progress_stats;
+mod section;
 mod segment;
 mod symbol;
 mod symbol_comparison_info;
 mod symbol_decomp_state;
 pub mod utils;
 
-pub use file::{File, PathDecompSettings};
 pub use found_symbol_info::FoundSymbolInfo;
 pub use mapfile::MapFile;
 pub use maps_comparison_info::MapsComparisonInfo;
 pub use progress_stats::ProgressStats;
+pub use section::{PathDecompSettings, Section};
 pub use segment::Segment;
 pub use symbol::Symbol;
 pub use symbol_comparison_info::SymbolComparisonInfo;
 pub use symbol_decomp_state::{SymbolDecompState, SymbolDecompStateIter};
+
+// Renamed types
+#[deprecated(since = "2.8.0", note = "Use `Section` instead")]
+pub use Section as File;
 
 #[macro_use]
 extern crate lazy_static;
@@ -33,7 +37,7 @@ use pyo3::prelude::*;
 fn mapfile_parser(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<mapfile::MapFile>()?;
     m.add_class::<segment::Segment>()?;
-    m.add_class::<file::File>()?;
+    m.add_class::<section::Section>()?;
     m.add_class::<symbol::Symbol>()?;
     m.add_class::<found_symbol_info::python_bindings::PyFoundSymbolInfo>()?;
     m.add_class::<symbol_comparison_info::python_bindings::PySymbolComparisonInfo>()?;